+++ src/datascript/query.cljs
@@ -23,6 +23,7 @@
 (defn intersect-keys [attrs1 attrs2]
   (set/intersection (set (keys attrs1))
                     (set (keys attrs2))))
+
 (defn concatv [& xs]
   (vec (apply concat xs)))